backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
pillow==10.1.0
imagehash==4.3.1
sentence-transformers==2.2.2
nltk==3.8.1
textstat==0.7.3
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
praw==7.7.1
tweepy==4.14.0
google-api-python-client==2.108.0
python-dateutil==2.8.2
pytz==2023.3

